Tuned fourier filtering (see notebook)
Extraction of the high-frequency component works when it has a small
amplitude (which is true along the T axis in the temperature progression
xrd dataset). In q, the peaks can't be separated from the background in
a self-consistent way, so we instead use an ad-hoc procedure to identify
the peak positions, mask those points out, and interpolate the
background.